Refactor app to mvc
