+++ weather-app/app/build.gradle
@@ -31,4 +31,5 @@ dependencies {
     implementation 'com.android.support:design:27.+'
     implementation 'com.android.support:support-v4:27.+'
     testImplementation 'junit:junit:4.12'
+    testImplementation 'org.mockito:mockito-core:1.10.19'
 }
