+++ weather-app/app/src/main/java/com/rperazzo/weatherapp/model/settings/SettingsRepository.java
@@ -1,0 +1,6 @@
+package com.rperazzo.weatherapp.model.settings;
+
+public interface SettingsRepository {
+    void setTemperatureUnit(String value);
+    String getTemperatureUnit();
+}
