+++ weather-app/app/src/main/java/com/rperazzo/weatherapp/model/settings/local/SettingsLocal.java
@@ -1,0 +1,6 @@
+package com.rperazzo.weatherapp.model.settings.local;
+
+public interface SettingsLocal {
+    void setTemperatureUnit(String value);
+    String getTemperatureUnit();
+}
