+++ weather-app/app/src/main/java/com/rperazzo/weatherapp/model/weather/WeatherRepository.java
@@ -1,0 +1,5 @@
+package com.rperazzo.weatherapp.model.weather;
+
+public interface WeatherRepository {
+    void search(String text, String units);
+}
