+++ weather-app/app/src/main/java/com/rperazzo/weatherapp/model/weather/remote/FindResult.java
@@ -1,0 +1,13 @@
+package com.rperazzo.weatherapp.model.weather.remote;
+
+import com.rperazzo.weatherapp.model.weather.City;
+
+import java.util.List;
+
+public class FindResult {
+    public final List<City> list;
+
+    public FindResult(List<City> list) {
+        this.list = list;
+    }
+}
