+++ weather-app/app/src/main/java/com/rperazzo/weatherapp/model/weather/remote/WeatherManager.java
@@ -1,0 +1,27 @@
+package com.rperazzo.weatherapp.model.weather.remote;
+
+import okhttp3.OkHttpClient;
+import retrofit2.Retrofit;
+import retrofit2.converter.gson.GsonConverterFactory;
+
+public class WeatherManager {
+
+    private static final String API_URL =
+            "http://api.openweathermap.org/data/2.5/";
+    public static final String API_KEY =
+            "520d6b47a12735bee8f69c57737d145f";
+
+    private static OkHttpClient mClient = new OkHttpClient();
+
+    public static WeatherService getService() {
+        Retrofit retrofit = new Retrofit.Builder()
+                .baseUrl(API_URL)
+                .addConverterFactory(GsonConverterFactory.create())
+                .client(mClient)
+                .build();
+
+        return retrofit.create(WeatherService.class);
+    }
+}
+
+
