+++ weather-app/app/src/main/java/com/rperazzo/weatherapp/model/weather/remote/WeatherRemote.java
@@ -1,0 +1,7 @@
+package com.rperazzo.weatherapp.model.weather.remote;
+
+import com.rperazzo.weatherapp.ui.WeatherView;
+
+public interface WeatherRemote {
+    void find(String text, String units, final WeatherView view);
+}
