+++ weather-app/app/src/main/java/com/rperazzo/weatherapp/model/weather/remote/WeatherService.java
@@ -1,0 +1,14 @@
+package com.rperazzo.weatherapp.model.weather.remote;
+
+import retrofit2.Call;
+import retrofit2.http.GET;
+import retrofit2.http.Query;
+
+public interface WeatherService {
+    @GET("find")
+    Call<FindResult> find(
+            @Query("q") String cityName,
+            @Query("units") String units,
+            @Query("appid") String apiKey
+    );
+}
