+++ weather-app/app/src/main/java/com/rperazzo/weatherapp/ui/WeatherView.java
@@ -1,0 +1,10 @@
+package com.rperazzo.weatherapp.ui;
+
+import com.rperazzo.weatherapp.model.weather.City;
+
+import java.util.List;
+
+public interface WeatherView {
+    void onFinishLoading(List<City> list);
+    void onFinishLoadingWithError(String error);
+}
